Avoid using viper.IsSet, which is broken
In some places we need to know whether a flag has been explicitly
set. When it was just flags, `pflag.Changed` served this purpose. But
`viper.IsSet`, which is supposed to do the same thing over flags,
config files, and environment entries, is broken:
https://github.com/spf13/viper/issues/276.

Therefore: I've made my own isSet predicate, which checks the flags,
config and environment individually. Since we can't actually check the
environment separately, it has to construct a viper value that _only_
looks at the env, and check that.